// autogpt_platform/frontend/src/app/api/__generated__/models/docPageResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ResponseType } from "./responseType";
import type { DocPageResponseSessionId } from "./docPageResponseSessionId";
import type { DocPageResponseDocUrl } from "./docPageResponseDocUrl";

/**
 * Response for get_doc_page tool.
 */
export interface DocPageResponse {
  type?: ResponseType;
  message: string;
  session_id?: DocPageResponseSessionId;
  title: string;
  path: string;
  content: string;
  doc_url?: DocPageResponseDocUrl;
}