// autogpt_platform/frontend/src/app/api/__generated__/models/docSearchResultsResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ResponseType } from "./responseType";
import type { DocSearchResultsResponseSessionId } from "./docSearchResultsResponseSessionId";
import type { DocSearchResult } from "./docSearchResult";

/**
 * Response for search_docs tool.
 */
export interface DocSearchResultsResponse {
  type?: ResponseType;
  message: string;
  session_id?: DocSearchResultsResponseSessionId;
  results: DocSearchResult[];
  count: number;
  query: string;
}