// autogpt_platform/frontend/src/app/api/__generated__/models/noResultsResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ResponseType } from "./responseType";
import type { NoResultsResponseSessionId } from "./noResultsResponseSessionId";

/**
 * Response when no agents found.
 */
export interface NoResultsResponse {
  type?: ResponseType;
  message: string;
  session_id?: NoResultsResponseSessionId;
  suggestions?: string[];
  name?: string;
}